# swagger.yaml
openapi: 3.0.0
info:
  title: Ticketing Platform API
  description: API for managing events, tickets, and user authentication
  version: 1.0.0
  contact:
    name: SmartTicket Support
servers:
  - url: http://localhost:${PORT}/api
    description: Development server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [regular, admin]
    Event:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        description:
          type: string
        date:
          type: string
          format: date-time
        venue_id:
          type: integer
          format: int64
        organizer_id:
          type: integer
          format: int64
        event_type_id:
          type: integer
          format: int64
        total_tickets:
          type: integer
        available_tickets:
          type: integer
        price:
          type: number
          format: float
        venue:
          $ref: '#/components/schemas/Venue'
        organizer:
          $ref: '#/components/schemas/Organizer'
        eventType:
          $ref: '#/components/schemas/EventType'
        artists:
          type: array
          items:
            $ref: '#/components/schemas/Artist'
    Venue:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        address:
          type: string
        city_id:
          type: integer
          format: int64
        capacity:
          type: integer
        description:
          type: string
          nullable: true
    Organizer:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        email:
          type: string
        bio:
          type: string
          nullable: true
    EventType:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
          nullable: true
    Artist:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        genre:
          type: string
        bio:
          type: string
          nullable: true
...
    Error:
      type: object
      properties:
        message:
          type: string
        error:
          type: string
paths:
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Register a new user
      description: Create a new user account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - email
                - password
              properties:
                username:
                  type: string
                  example: johndoe
                email:
                  type: string
                  format: email
                  example: john@example.com
                password:
                  type: string
                  format: password
                  example: password123
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  username:
                    type: string
                  email:
                    type: string
        '400':
          description: User already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/login:
    post:
      tags:
        - Authentication
      summary: User login
      description: Authenticate a user and get an access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: john@example.com
                password:
                  type: string
                  format: password
                  example: password123
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
        '400':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/me:
    get:
      tags:
        - Authentication
      summary: Get basic user information
      description: Get basic information about the currently authenticated user
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User information retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/profile:
    get:
      tags:
        - Authentication
      summary: Get complete user profile
      description: Get comprehensive profile information for the currently authenticated user
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User profile retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  username:
                    type: string
                  email:
                    type: string
                  role:
                    type: string
                  is_active:
                    type: boolean
                  created_at:
                    type: string
                    format: date-time
                  updated_at:
                    type: string
                    format: date-time
        '401':
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /tickets/buy:
    post:
      tags:
        - Tickets
      summary: Purchase a ticket
      description: Buy a ticket for an event (requires authentication)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - event_id
              properties:
                event_id:
                  type: integer
                  description: ID of the event to purchase a ticket for
                  example: 1
      responses:
        '201':
          description: Ticket purchased successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Ticket acheté avec succès"
                  ticket:
                    type: object
                    properties:
                      id:
                        type: integer
                      event_id:
                        type: integer
                      user_id:
                        type: integer
                      price_paid:
                        type: number
                      status:
                        type: string
                      ticket_code:
                        type: string
                      purchase_date:
                        type: string
                        format: date-time
        '400':
          description: Invalid request or no tickets available
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized - Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Event not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /tickets/user:
    get:
      tags:
        - Tickets
      summary: Get current user's tickets (alternative endpoint)
      description: Retrieve all tickets purchased by the currently authenticated user
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User's tickets retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    event_id:
                      type: integer
                    user_id:
                      type: integer
                    price_paid:
                      type: number
                    purchase_date:
                      type: string
                      format: date-time
                    status:
                      type: string
                    ticket_code:
                      type: string
                    event:
                      $ref: '#/components/schemas/Event'
        '401':
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /tickets/{id}:
    get:
      tags:
        - Tickets
      summary: Get ticket by ID
      description: Retrieve a specific ticket by its ID (requires authentication)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Ticket ID
          schema:
            type: integer
      responses:
        '200':
          description: Ticket retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  event_id:
                    type: integer
                  user_id:
                    type: integer
                  price_paid:
                    type: number
                  purchase_date:
                    type: string
                    format: date-time
                  status:
                    type: string
                  ticket_code:
                    type: string
                  event:
                    $ref: '#/components/schemas/Event'
                  user:
                    type: object
                    properties:
                      id:
                        type: integer
                      username:
                        type: string
                      email:
                        type: string
        '401':
          description: Unauthorized - Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - Not authorized to access this ticket
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Ticket not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'